% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_elec}
\alias{ny_elec}
\title{Net Generation of Electricity for the New York Region}
\format{A tsibble object with hourly intervals}
\source{
US Energy Information Administration (Jan 2020) \href{https://www.eia.gov/}{website}
}
\usage{
ny_elec
}
\description{
The total hourly generation of electricity in the New York region since July 2015

Units: megawatthours

Time zone: UTC
}
\details{
The dataset contains the hourly generation of electricity in the New York region in megawatthours.
The series index is in UTC time zone
}
\examples{

data(ny_elec)

library(plotly)

plot_ly(data = ny_elec,
        x = ~ date_time,
        y = ~ y,
        type = "scatter",
        mode = "lines")

}
\keyword{datasets}
\keyword{electricity}
\keyword{new-york}
\keyword{time-series}
