% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_gas}
\alias{ny_gas}
\title{The New York Natural Gas Residential Monthly Consumption (1997-2019)}
\format{
A tsibble object with monthly intervals
}
\source{
US Energy Information Administration (Jan 2020) \href{https://www.eia.gov/}{website}
}
\usage{
ny_gas
}
\description{
Units: Million Cubic Feet
}
\details{
Time zone: UTC

The dataset contains the monthly consumption of natural gas in the New York in million cubic feet.
The series index is in year-month format
}
\examples{

data(ny_gas)

library(plotly)

plot_ly(data = ny_gas,
        x = ~ date,
        y = ~ y,
        type = "scatter",
        mode = "lines")

}
\keyword{datasets}
\keyword{gas}
\keyword{natural}
\keyword{new-york}
\keyword{time-series}
