% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_fc}
\alias{plot_fc}
\title{Plotting the forecast output}
\usage{
plot_fc(forecast, theme = "normal")
}
\arguments{
\item{forecast}{A forecastLM object}

\item{theme}{A character, defines the color theme to be used in the plot output.
Available themes - "normal" (default), "darkBlue", "darkPink", "darkGreen", "classic", "lightBeige"}
}
\value{
A plotly object
}
\description{
Plotting the forecast output
}
\examples{

# Load the data
data(ny_gas)

head(ny_gas)

# Train a time series forecasting model
md <- trainLM(input = ny_gas,
              y = "y",
              trend = list(linear = TRUE),
              seasonal = "month",
              lags = c(1, 12))

fc <- forecastLM(model = md, h = 60)

# Plot the forecast model
plot_fc(fc)

# Use different plot theme
plot_fc(fc, theme = "darkPink")

}
