% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_fit}
\alias{plot_fit}
\title{Plotting the Fitted Values vs. Actuals}
\usage{
plot_fit(model)
}
\arguments{
\item{model}{A trainLM object}
}
\description{
Plotting the model's fitted values against the series
}
\examples{

data(ny_gas)

head(ny_gas)

# Training a model
md <- trainLM(input = ny_gas,
              y = "y",
              trend = list(linear = TRUE),
              seasonal = "month",
              lags = c(1, 12))

# plot fitted values
plot_fit(md)

}
