% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_res}
\alias{plot_res}
\title{Plotting the residuals}
\usage{
plot_res(model, na.rm = FALSE, margin = 0.04)
}
\arguments{
\item{model}{A trainLM object}

\item{na.rm}{A boolean, if TRUE will ignore missing values, by default set to FALSE}

\item{margin}{A numeric, define the margin space between the subplot components}
}
\description{
Plotting the residuals
}
\examples{

# Train a time series forecasting model
md <- trainLM(input = AirPassengers,
              trend = list(log = TRUE, linear = TRUE),
              seasonal = "month")

# create a residuals plot
plot_res(md)

}
