% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_functions.R
\name{trainLM}
\alias{trainLM}
\title{Training a Forecasting Model with Linear Regression Framework}
\usage{
trainLM(input, y = NULL, x = NULL, seasonal = NULL,
  trend = list(linear = TRUE, exponential = FALSE, log = FALSE, power =
  FALSE), lags = NULL, events = NULL, scale = NULL, step = FALSE,
  ...)
}
\arguments{
\item{input}{A tsibble or ts object}

\item{y}{A character, the column name of the depended variable of the input object, required (and applicable) only when the input is tsibble object}

\item{x}{A character, the column names of the independent variable of the input object, applicable when using tsibble object with regressors}

\item{seasonal}{A character, optional, create categorical variable/s to model the single or multiple seasonal components. Supporting the
following frequencies structure:

`quarter` - for quarterly frequency

`month` - for monthly frequency

`week` - for weekly frequency

`yday` - for daily frequency

`wday` - for day of the week frequency

`hour` - for hourly frequency

The argument supports multiple seasonality cases such as daily and day of the week effect}

\item{trend}{A list, define the trend structure. Possible arguments -

`linear`, a boolean variable, if set to TRUE defines a linear trend (e.g., index of 1,2,3,...,t, for a series with t observations)

`power``, a numeric value, defines the polynomial degree of the series index (for example, a power = 0.5 define a square root index and power = 2 represents a squared index).
By default set to NULL

`exponential`, a boolean variable, if set to TRUE defines an exponential trend

`log`, a boolean variable, if set to TRUE defines a log transformation for the trend.

By default, the `trend` argument is set to a linear trend (i.e., power = 1)}

\item{lags}{A positive integer, defines the series lags to be used as input to the model (equivalent to AR process)}

\item{events}{A list, an optional, create hot encoding variables based on date/time objects,
where the date/time objects must align with the input object index class (may not work when the input object is 'ts')}

\item{scale}{A character, scaling options of the series, methods available -
c("log", "normal", "standard") for log transformation, normalization, or standardization of the series, respectively.
If set to NULL (default), no transformation will occur}

\item{step}{A boolean, if set to TRUE will use apply the stepwise function for variable selection using the \code{\link[stats]{step}} function}

\item{...}{Optional, a list of arguments to pass to the \code{\link[stats]{step}} function}
}
\description{
Methods for forecasting regular time series data based on a linear regression model
}
\details{
The trainLM function provides a flexible framework for forecasting regular time-series data with the linear regression model with the \code{\link[stats]{lm}} function.
The function arguments enable a fast extraction and generation of new features from the input series, such as seasonal, trend, lags, outliers, and special events.
}
